[refactor] : modal 기능 분리 및 일괄 적용
