+++ app/signin/page.tsx
@@ -176,7 +176,7 @@ export default function Page() {
           </div>
         </form>
       </div>
-      {isModalOpen && <Modal type="big" message={modalMessage} onClose={handleCloseModal} />}
+      {isModalOpen && <Modal message={modalMessage} onClose={handleCloseModal} />}
     </>
   );
 }
