+++ components/reservation-calendar/big-calendar.tsx
@@ -2,8 +2,8 @@ import React, {useState, useEffect, useRef} from 'react';
 import {Calendar} from 'antd';
 import type {Dayjs} from 'dayjs';
 import CalendarHeader from './calendar-header';
-import ReservationContainer from '../common/modal/reservation-container';
-import ReservationModal from '../common/modal/reservation-modal';
+import ReservationContainer from './reservation-container';
+import ReservationModal from './reservation-modal';
 import {calendarStatusLabels} from '@/constant/reservation-list-constant';
 import dayjs from 'dayjs';
 import updateLocale from 'dayjs/plugin/updateLocale';
