+++ components/reservation-list/reservation-list.tsx
@@ -1,21 +1,21 @@
 'use client';
 import React, {useState, useEffect} from 'react';
-import Image from 'next/image';
-import Button from '@/components/common/button';
-import Modal from '@/components/common/modal/modal';
-import ReviewModal from '@/components/common/modal/review-modal';
-import CustomSelect from '@/components/reservation-list/custom-select';
-import {statusLabels, buttonByStatus} from '@/constant/reservation-list-constant';
-import NonDataPage from '../common/non-data';
-import closeButton from '@/public/icon/ic_close_button.svg';
-import {getReservationList} from '@/service/api/reservation-list/getReservation.api';
-import {useInfiniteQuery} from '@tanstack/react-query';
-import {ReservationListResponse} from '@/types/reservation-list';
 import {ScaleLoader} from 'react-spinners';
-import FormattedPrice from '@/utils/formatted-price';
 import {useInView} from 'react-intersection-observer';
-import FormatDate from '@/utils/format-date';
+import {useInfiniteQuery} from '@tanstack/react-query';
+import Image from 'next/image';
 import {useRouter} from 'next/navigation';
+import Button from '@/components/common/button';
+import ReviewModal from '@/components/reservation-list/review-modal';
+import CustomSelect from '@/components/reservation-list/custom-select';
+import NonDataPage from '@/components/common/non-data';
+import ReservationModal from '@/components/reservation-list/reservation-modal';
+import {statusLabels, buttonByStatus} from '@/constant/reservation-list-constant';
+import {getReservationList} from '@/service/api/reservation-list/getReservation.api';
+import {ReservationListResponse} from '@/types/reservation-list';
+import FormatDate from '@/utils/format-date';
+import FormattedPrice from '@/utils/formatted-price';
+import closeButton from '@/public/icon/ic_close_button.svg';
 
 export const statusLabelsColor: Record<string, string> = {
   pending: 'text-blue-100',
@@ -76,7 +76,7 @@ export default function ReservationList() {
   const getModalContent = () => {
     switch (modalType) {
       case 'pending':
-        return <Modal reservationId={selectedId} type="small" message="예약을 취소하시겠습니까?" onClose={() => setIsOpen(false)} />;
+        return <ReservationModal reservationId={selectedId} message="예약을 취소하시겠습니까?" onClose={() => setIsOpen(false)} />;
       case 'completed':
         const selectedData = reservationList.find(reservation => reservation.status === 'completed' && reservation.id === selectedId);
         return <ReviewModal data={selectedData} message={'후기 작성'} onClose={() => setIsOpen(false)} />;
